+++ app/core/Engine/Engine.ts
@@ -1061,6 +1061,7 @@ export class Engine {
         name: 'BridgeStatusController',
         allowedActions: [
           'AccountsController:getSelectedAccount',
+          'AccountsController:getSelectedMultichainAccount',
           'NetworkController:getNetworkClientById',
           'NetworkController:findNetworkClientIdByChainId',
           'NetworkController:getState',
@@ -1068,8 +1069,9 @@ export class Engine {
         ],
         allowedEvents: [],
       }),
+      state: initialState.BridgeStatusController,
       clientId: BridgeClientId.MOBILE,
-      fetchFn: fetch,
+      fetchFn: handleFetch,
     });
 
     const existingControllersByName = {